jaeger refactor; minor updates
